+++ uamqp/__init__.py
@@ -32,7 +32,7 @@ except (SyntaxError, ImportError):
     pass  # Async not supported.
 
 
-__version__ = "1.1.0rc1"
+__version__ = "1.1.0rc2"
 
 
 _logger = logging.getLogger(__name__)