+++ uamqp/receiver.py
@@ -146,6 +146,8 @@ class MessageReceiver(object):
             if _previous_state == constants.MessageReceiverState.Opening \
                     and _new_state == constants.MessageReceiverState.Error:
                 _logger.info("Receiver link failed to open - expecting to receive DETACH frame.")
+            elif self._session._link_error:  # pylint: disable=protected-access
+                _logger.info("Receiver link ATTACH frame invalid - expecting to receive DETACH frame.")
             else:
                 self.on_state_changed(_previous_state, _new_state)
         except KeyboardInterrupt: